Updated to new version of scipy; removed all pyc files

+++ app/src/Version_2.1/sessionCalibrationProcess/scipy/io/tests/test_wavfile.py
@@ -1,28 +1,28 @@
 from __future__ import division, print_function, absolute_import
 
 import os
+import sys
 import tempfile
 import warnings
+from io import BytesIO
 
 import numpy as np
-from numpy.testing import assert_equal, assert_, assert_raises, assert_array_equal
-from numpy.testing.utils import WarningManager
+from numpy.testing import (assert_equal, assert_, assert_raises,
+    assert_array_equal, run_module_suite)
+
 from scipy.io import wavfile
 
 
 def datafile(fn):
     return os.path.join(os.path.dirname(__file__), 'data', fn)
 
+
 def test_read_1():
     for mmap in [False, True]:
-        warn_ctx = WarningManager()
-        warn_ctx.__enter__()
-        try:
+        with warnings.catch_warnings():
             warnings.simplefilter('ignore', wavfile.WavFileWarning)
-            rate, data = wavfile.read(datafile('test-44100-le-1ch-4bytes.wav'),
+            rate, data = wavfile.read(datafile('test-44100Hz-le-1ch-4bytes.wav'),
                                       mmap=mmap)
-        finally:
-            warn_ctx.__exit__()
 
         assert_equal(rate, 44100)
         assert_(np.issubdtype(data.dtype, np.int32))
@@ -30,9 +30,10 @@ def test_read_1():
 
         del data
 
+
 def test_read_2():
     for mmap in [False, True]:
-        rate, data = wavfile.read(datafile('test-8000-le-2ch-1byteu.wav'),
+        rate, data = wavfile.read(datafile('test-8000Hz-le-2ch-1byteu.wav'),
                                   mmap=mmap)
         assert_equal(rate, 8000)
         assert_(np.issubdtype(data.dtype, np.uint8))
@@ -40,21 +41,65 @@ def test_read_2():
 
         del data
 
+def test_read_3():
+    for mmap in [False, True]:
+        rate, data = wavfile.read(datafile('test-44100Hz-2ch-32bit-float-le.wav'),
+                                  mmap=mmap)
+        assert_equal(rate, 44100)
+        assert_(np.issubdtype(data.dtype, np.float32))
+        assert_equal(data.shape, (441, 2))
+
+        del data
+
+def test_read_4():
+    for mmap in [False, True]:
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore', wavfile.WavFileWarning)
+            rate, data = wavfile.read(datafile('test-48000Hz-2ch-64bit-float-le-wavex.wav'),
+                                      mmap=mmap)
+
+        assert_equal(rate, 48000)
+        assert_(np.issubdtype(data.dtype, np.float64))
+        assert_equal(data.shape, (480, 2))
+
+        del data
+
+
+def test_read_5():
+    for mmap in [False, True]:
+        rate, data = wavfile.read(datafile('test-44100Hz-2ch-32bit-float-be.wav'),
+                                  mmap=mmap)
+        assert_equal(rate, 44100)
+        assert_(np.issubdtype(data.dtype, np.float32))
+        assert_(data.dtype.byteorder == '>' or (sys.byteorder == 'big' and
+                                                data.dtype.byteorder == '='))
+        assert_equal(data.shape, (441, 2))
+
+        del data
+
+
 def test_read_fail():
     for mmap in [False, True]:
         fp = open(datafile('example_1.nc'))
         assert_raises(ValueError, wavfile.read, fp, mmap=mmap)
         fp.close()
 
-def _check_roundtrip(rate, dtype, channels):
-    fd, tmpfile = tempfile.mkstemp(suffix='.wav')
-    try:
-        os.close(fd)
 
+def _check_roundtrip(realfile, rate, dtype, channels):
+    if realfile:
+        fd, tmpfile = tempfile.mkstemp(suffix='.wav')
+        os.close(fd)
+    else:
+        tmpfile = BytesIO()
+    try:
         data = np.random.rand(100, channels)
         if channels == 1:
             data = data[:,0]
-        data = (data*128).astype(dtype)
+        if dtype.kind == 'f':
+            # The range of the float type should be in [-1, 1]
+            data = data.astype(dtype)
+        else:
+            data = (data*128).astype(dtype)
 
         wavfile.write(tmpfile, rate, data)
 
@@ -67,18 +112,38 @@ def _check_roundtrip(rate, dtype, channels):
 
             del data2
     finally:
-        os.unlink(tmpfile)
+        if realfile:
+            os.unlink(tmpfile)
+
 
 def test_write_roundtrip():
-    for signed in ('i', 'u'):
-        for size in (1, 2, 4, 8):
-            if size == 1 and signed == 'i':
-                # signed 8-bit integer PCM is not allowed
-                continue
-            for endianness in ('>', '<'):
-                if size == 1 and endianness == '<':
+    for realfile in (False, True):
+        for dtypechar in ('i', 'u', 'f', 'g', 'q'):
+            for size in (1, 2, 4, 8):
+                if size == 1 and dtypechar == 'i':
+                    # signed 8-bit integer PCM is not allowed
                     continue
-                for rate in (8000, 32000):
-                    for channels in (1, 2, 5):
-                        dt = np.dtype('%s%s%d' % (endianness, signed, size))
-                        yield _check_roundtrip, rate, dt, channels
+                if size > 1 and dtypechar == 'u':
+                    # unsigned > 8-bit integer PCM is not allowed
+                    continue
+                if (size == 1 or size == 2) and dtypechar == 'f':
+                    # 8- or 16-bit float PCM is not expected
+                    continue
+                if dtypechar in 'gq':
+                    # no size allowed for these types
+                    if size == 1:
+                        size = ''
+                    else:
+                        continue
+
+                for endianness in ('>', '<'):
+                    if size == 1 and endianness == '<':
+                        continue
+                    for rate in (8000, 32000):
+                        for channels in (1, 2, 5):
+                            dt = np.dtype('%s%s%s' % (endianness, dtypechar, size))
+                            yield _check_roundtrip, realfile, rate, dt, channels
+
+
+if __name__ == "__main__":
+    run_module_suite()
